% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{BinaryEncode}
\alias{BinaryEncode}
\title{Binary encoding of categorical variables}
\usage{
BinaryEncode(c, m)
}
\arguments{
\item{c:}{observed categorical variables in discrete space}

\item{m:}{number of elements in such discrete space.}
}
\description{
Binary encoding of categorical variables
}
