% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Branin}
\alias{Branin}
\title{Annotated Branin Function for Nominal Case Simulation}
\usage{
Branin(
  x1,
  x2,
  x1.ub = 2,
  x1.lb = -2,
  x2.ub = 2,
  x2.lb = -2,
  a = 1,
  b = 5.1/(4 * pi^2),
  c = 5/pi,
  s = 10,
  r = 6,
  t = 1/(8 * pi)
)
}
\arguments{
\item{u1, u2}{Latent continuous coordinates.}
}
\value{
A list with response 'f' and category 'c'.
}
\description{
Acts as the oracle for the simulation study in Section 5.2.
}
\details{
Maps U to 5 regions:
Region 1-2: Separated by linear boundaries.
Region 3-5: Partitioned via Voronoi diagram based on Branin's global minima.
}
