% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{Classify.stickbreak}
\alias{Classify.stickbreak}
\title{Compute Categorical Probabilities via Stick-Breaking Multinomial Logistic
Regression}
\usage{
Classify.stickbreak(gamma.mat, U)
}
\arguments{
\item{gamma.mat:}{coefficient matrix within intercept}

\item{U:}{matrix of continues latent variable.}
}
\value{
A list contain: prob,prob.tilde,linear score psi.
}
\description{
Compute Categorical Probabilities via Stick-Breaking Multinomial Logistic
Regression
}
