% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{conditional_mean_scores}
\alias{conditional_mean_scores}
\title{Conditional mean scores for ordinal classes}
\usage{
conditional_mean_scores(c_ord, m)
}
\arguments{
\item{c_ord}{Integer vector of class labels (1 … m).}

\item{m}{Number of classes.}
}
\value{
Numeric vector of length \code{m} giving the conditional
mean for each class.
}
\description{
Computes the expected value of a standard normal random variable
truncated to each ordinal class based on empirical quantiles
of the class labels.
}
