% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{DrawTrunNormal}
\alias{DrawTrunNormal}
\title{Simulate Independent Truncated Normal Variables by Inverse Probability Integral Transform.}
\usage{
DrawTrunNormal(mu, sigma, lb, ub)
}
\arguments{
\item{mu}{Vector of means for the underlying normal distributions.}

\item{sigma}{Vector of standard deviations.}

\item{lb}{Vector of lower bounds for truncation.}

\item{ub}{Vector of upper bounds for truncation.}
}
\value{
A vector of samples from the truncated normal distribution(s).
}
\description{
Simulate Independent Truncated Normal Variables by Inverse Probability Integral Transform.
}
