% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{EnergyDistance}
\alias{EnergyDistance}
\title{Compute energy distance.}
\usage{
EnergyDistance(X, Y)
}
\arguments{
\item{X:}{predictions (a vector);}

\item{Y:}{samples from the target}
}
\description{
Compute energy distance.
}
