% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{EstimateEvidence}
\alias{EstimateEvidence}
\title{Given hyper-parameter (rho, theta), estimate the marginal likelihood up to a constant factor by Monte Carlo.}
\usage{
EstimateEvidence(theta, rho, U.hist, IG.a.error, IG.b.error, X, y, d = 1)
}
\arguments{
\item{theta}{lengthscales of x and u}

\item{rho}{signal-to-noise ratio. That is, the GP variance is rho^2*sigma.error2}

\item{U.hist}{Samples of U from p(U|X, c)}

\item{IG.a.error, IG.b.error}{parameters of the IG prior for sigma.error2}

\item{d:}{the dimension of u}
}
\value{
A vector storing the samples of marginal likelihood.
}
\description{
Given hyper-parameter (rho, theta), estimate the marginal likelihood up to a constant factor by Monte Carlo.
}
