% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{get.initial.nominal}
\alias{get.initial.nominal}
\title{Generate initial values for Ordinal Models}
\usage{
get.initial.nominal(Gibbs.data)
}
\arguments{
\item{Gibbs.data}{List containing data parameters (n, m, c, U.obs, u.obs.idx, etc.).}
}
\value{
A list of initial values (e.g., U0, Tau0).
}
\description{
Generate initial values for Ordinal Models
}
