% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{get.initial}
\alias{get.initial}
\title{Unified Function to Generate Initial Values for U (and Tau)}
\usage{
get.initial(input.type, Gibbs.data = list())
}
\arguments{
\item{input.type}{String. "ordinal" or "nominal".}

\item{Gibbs.data}{List containing data parameters (n, m, c, U.obs, u.obs.idx, etc.).}
}
\value{
A list of initial values (e.g., U0, Tau0).
}
\description{
Unified Function to Generate Initial Values for U (and Tau)
}
