% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{Gibbs.EB.ordinal}
\alias{Gibbs.EB.ordinal}
\title{Empirical Bayes for Gibbs for tuning, Ordinal case}
\usage{
Gibbs.EB.ordinal(Gibbs.param, Gibbs.data, Gibbs.hyper, Gibbs.other)
}
\arguments{
\item{Gibbs.param}{see above}

\item{Gibbs.data}{see above}

\item{Gibbs.hyper}{see above}

\item{Gibbs.other}{see above}
}
\value{
related list (for detail, see above)
}
\description{
Empirical Bayes for Gibbs for tuning, Ordinal case
}
