% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{Gibbs.nominal}
\alias{Gibbs.nominal}
\title{Gibbs sampler for nominal case (Note that well organized method now)}
\usage{
Gibbs.nominal(Gibbs.param, Gibbs.data, Gibbs.hyper, GP.param, Gibbs.other)
}
\arguments{
\item{Gibbs.param:}{see above}

\item{Gibbs.data:}{see above}

\item{Gibbs.hyper:}{see above}

\item{GP.param:}{see above}

\item{Gibbs.other:}{see above}
}
\value{
see above
}
\description{
Gibbs sampler for nominal case (Note that well organized method now)
}
