% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{Gibbs.ordinal}
\alias{Gibbs.ordinal}
\title{Gibbs sampler for ordinal case}
\usage{
Gibbs.ordinal(Gibbs.param, Gibbs.data, Gibbs.hyper, GP.param, Gibbs.other)
}
\arguments{
\item{Gibbs.param:}{see above}

\item{Gibbs.data:}{see above}

\item{Gibbs.hyper:}{see above}

\item{GP.param:}{see above}

\item{Gibbs.other:}{see above}
}
\value{
see above
}
\description{
Gibbs sampler for ordinal case
}
