% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{gp_mle_fit_learned_embedding}
\alias{gp_mle_fit_learned_embedding}
\title{Fit a GP with a learned ordinal embedding}
\usage{
gp_mle_fit_learned_embedding(x, c_ord, y, m, n_starts = 8L)
}
\arguments{
\item{x}{Numeric vector of continuous covariate.}

\item{c_ord}{Integer class labels.}

\item{y}{Numeric response vector.}

\item{m}{Number of classes.}

\item{n_starts}{Number of random starting values.}
}
\value{
A list with \code{par} (log-hyperparameters), \code{a} (log-ratios
for embedding), \code{z_scores} (final embedding), \code{value}, etc.
}
\description{
Estimates GP hyperparameters together with a learned monotone
embedding of the ordinal classes by maximum likelihood.
Multiple random starts are used to avoid local optima.
}
