% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{Gradient.nominal}
\alias{Gradient.nominal}
\title{Gradient of the Log Conditional Density w.r.t U (Nominal)}
\usage{
Gradient.nominal(U, phi, r, K, mn.u, Qn.u, length.scale.u)
}
\arguments{
\item{U}{Matrix of current latent variables (n x d).}

\item{phi}{Auxiliary matrix/vector from data augmentation.}

\item{r}{GP residual vector (C^-1 * y).}

\item{K}{Current kernel matrix.}

\item{mn.u}{Posterior mean for latent variables (n x d).}

\item{Qn.u}{Array of precision matrices (d x d x n) for the Gaussian part.}

\item{length.scale.u}{Length-scale parameter for latent input u.}
}
\value{
A matrix of gradients (n x d).
}
\description{
Gradient of the Log Conditional Density w.r.t U (Nominal)
}
