% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{Gradient.ordinal}
\alias{Gradient.ordinal}
\title{Gradient of the Log Conditional Density w.r.t U (Ordinal)}
\usage{
Gradient.ordinal(U, r, phi, K, length.scale.u, X.tilde, beta, sigma.u2)
}
\arguments{
\item{U}{Vector of current latent variables.}

\item{r}{GP residual vector (C^-1 * y).}

\item{phi}{Auxiliary variable for the determinant-free trick.}

\item{K}{Kernel matrix.}

\item{length.scale.u}{Length-scale parameter for latent input u.}

\item{X.tilde}{Matrix of predictors with intercept.}

\item{beta}{Coefficient vector for u ~ X.}

\item{sigma.u2}{Variance of the latent variable u.}
}
\value{
A vector of gradients (same length as U).
}
\description{
Gradient of the Log Conditional Density w.r.t U (Ordinal)
}
