% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{impute.u.nominal}
\alias{impute.u.nominal}
\title{Impute New Latent Variables U for Nominal Models}
\usage{
impute.u.nominal(Gibbs.data, Gibbs.hist, Gibbs.other)
}
\description{
Impute New Latent Variables U for Nominal Models
}
