% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{impute.u.ordinal}
\alias{impute.u.ordinal}
\title{Impute New Latent Variables U for Ordinal Models}
\usage{
impute.u.ordinal(Gibbs.data, Gibbs.hist, Gibbs.other)
}
\description{
Impute New Latent Variables U for Ordinal Models
}
