% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{KernelMatrix}
\alias{KernelMatrix}
\title{Compute the kernel matrix K_{WV} with the inverse exponential squared kernel.}
\usage{
KernelMatrix(W, Theta, V = NULL, eps = 0)
}
\arguments{
\item{W:}{The data matrix with each row being an observation.}

\item{Theta:}{The last one is the overall variance, and the rest are lengthscales.}

\item{V:}{The second data matrix, by default is the same as W.}
}
\description{
Compute the kernel matrix K_{WV} with the inverse exponential squared kernel.
}
