% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{load_or_install}
\alias{load_or_install}
\title{Load or install a package}
\usage{
load_or_install(pkg)
}
\arguments{
\item{pkg}{Character string, name of the package.}
}
\value{
Invisibly loads the package.
}
\description{
Checks whether a package is installed. If not, installs it.
Then loads the package with suppressed startup messages.
}
