% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{make_default_control}
\alias{make_default_control}
\title{Create default MCMC control parameters}
\usage{
make_default_control(n, n_mis, preset = c("fast", "balanced", "thorough"))
}
\arguments{
\item{n}{Integer, total number of observations.}

\item{n_mis}{Integer, number of unobserved (latent) \code{u} values.}

\item{preset}{Character, one of \code{"fast"}, \code{"balanced"} or
\code{"thorough"}.}
}
\value{
A list of control parameters.
}
\description{
Chooses sampling intensities (fraction of local updates, block-
wise elliptical slice sampling frequency, etc.) according to a
preset profile.
}
