% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Plot.Data}
\alias{Plot.Data}
\title{Plot Scatterplots of Observed and Latent Data}
\usage{
Plot.Data(c, X, y, U, u.obs.idx)
}
\arguments{
\item{c}{Vector of qualitative inputs (category labels).}

\item{X}{Matrix/Vector of quantitative inputs (assumed 1D for visualization).}

\item{y}{Vector of observed responses.}

\item{U}{Vector of true latent continuous variables.}

\item{u.obs.idx}{Indices of the samples where the latent variable U is observed.}
}
\value{
A list containing two ggplot2 objects:
\item{ggfig1}{Scatterplot of y vs (x, c). Corresponds to Figure 5(a).
Shows the data as seen by a standard mixed-input model.}
\item{ggfig2}{Scatterplot of y vs observed (x, u). Corresponds to Figure 5(b).
Shows the "expensive" ground truth structure used as anchors.}
}
\description{
Plot Scatterplots of Observed and Latent Data
}
