% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{PlotResponse}
\alias{PlotResponse}
\title{Visualize the 2D Response Surface f(x, u)}
\usage{
PlotResponse(code, nx = 100, lb.u = -2, ub.u = 2, lb.x = -2, ub.x = 2)
}
\arguments{
\item{lb.u, ub.u:}{Lower and upper bounds for the latent U.}

\item{lb.x, ub.x:}{Lower and upper bounds for the X.}

\item{code:}{Character string indicating the test function to visualize
(e.g., 'sin', 'cos', 'quad', 'doppler').}

\item{nx:}{Resolution of the grid (number of points along each axis).}
}
\value{
A 3D perspective plot rendered using the \code{plot3D::persp3D} function.
}
\description{
Visualize the 2D Response Surface f(x, u)
}
