% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{predict.oracle.ordinal}
\alias{predict.oracle.ordinal}
\title{Oracle prediction using the data-generating process.}
\usage{
\method{predict}{oracle.ordinal}(Gibbs.data, Gibbs.param, Gibbs.other)
}
\arguments{
\item{Gibbs.data}{List containing data inputs (e.g., X, c, X.star, c.star, code, limits).}

\item{Gibbs.param}{List containing the true underlying parameters (e.g., beta, B.tilde, Tau,f).}

\item{Gibbs.other}{List containing other settings (e.g., uniform, n.mc, seed).}
}
\value{
A list containing the oracle histories (U.star.hist, y.predict.hist, etc.).
}
\description{
Oracle prediction using the data-generating process.
}
