% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{ResponseFunction}
\alias{ResponseFunction}
\title{Ground Truth Response Functions for Numerical Experiments}
\usage{
ResponseFunction(X, U, code)
}
\arguments{
\item{X:}{Matrix of observed quantitative inputs.}

\item{U:}{Matrix of latent continuous variables.}

\item{code:}{Character string defining the function type.}
}
\value{
A vector of noise-free response values.
}
\description{
Ground Truth Response Functions for Numerical Experiments
}
