% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{sample_gp_learned_embedding_predictive}
\alias{sample_gp_learned_embedding_predictive}
\title{Sample from the predictive distribution of the learned-embedding GP}
\usage{
sample_gp_learned_embedding_predictive(fit, x_star, c_star, n_draw = 1000)
}
\arguments{
\item{fit}{Output of \code{gp_mle_fit_learned_embedding}.}

\item{x_star}{Numeric vector of new \code{x} values.}

\item{c_star}{Integer vector of new class labels.}

\item{n_draw}{Number of draws.}
}
\value{
A matrix of predictive samples (n_draw x n_test).
}
\description{
Sample from the predictive distribution of the learned-embedding GP
}
