% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{solve_chol}
\alias{solve_chol}
\title{Solve linear system from Cholesky factor}
\usage{
solve_chol(U, b)
}
\arguments{
\item{U}{Upper triangular Cholesky factor.}

\item{b}{Right-hand side vector or matrix.}
}
\value{
Solution \code{x}.
}
\description{
Solves \verb{A x = b} given the upper triangular Cholesky factor
\code{U} of \code{A}.
}
