% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{StickBreak.inv}
\alias{StickBreak.inv}
\title{The mapping from m multinomial probabilities (adding up to 1) to (m-1) stick
-breaking probabilities}
\usage{
StickBreak.inv(prob)
}
\arguments{
\item{prob:}{an multinomial mass function on several points.}
}
\value{
m-1 Stick-breaking probabilities
}
\description{
The mapping from m multinomial probabilities (adding up to 1) to (m-1) stick
-breaking probabilities
}
