% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-helpers.R
\name{StickBreak}
\alias{StickBreak}
\title{The mapping from (m-1) stick-breaking probabilities to m multinomial
probabilities (adding up to 1).}
\usage{
StickBreak(prob.tilde)
}
\arguments{
\item{prob.tilde:}{m-1 Stick-breaking probabilities}
}
\description{
The mapping from (m-1) stick-breaking probabilities to m multinomial
probabilities (adding up to 1).
}
