% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{theta_logpost_1d}
\alias{theta_logpost_1d}
\title{Log-posterior for GP hyperparameters (1D)}
\usage{
theta_logpost_1d(
  y,
  u,
  Dx,
  logtheta,
  sigma2_eps,
  prior_mean = c(log(3), log(0.5), log(0.5)),
  prior_sd = c(1.5, 1.5, 1.5),
  bounds = rbind(log_rho = c(log(0.05), log(100)), log_theta_x = c(log(1e-04), log(100)),
    log_theta_u = c(log(1e-04), log(100)))
)
}
\arguments{
\item{y}{Response vector.}

\item{u}{Latent variable vector.}

\item{Dx}{Squared distance matrix.}

\item{logtheta}{Log hyperparameters.}

\item{sigma2_eps}{Error variance.}

\item{prior_mean}{Prior mean (optional).}

\item{prior_sd}{Prior sd (optional).}

\item{bounds}{Prior bounds (optional).}
}
\value{
Log posterior value.
}
\description{
Sum of log-prior and log-likelihood for the GP parameters.
By default uses the original prior constants.
}
