% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{update_tau_1d}
\alias{update_tau_1d}
\title{Gibbs update for thresholds (1D)}
\usage{
update_tau_1d(tau, u, c_ord, m, tau_bound = 8)
}
\arguments{
\item{tau}{Current threshold vector.}

\item{u}{Latent variable vector.}

\item{c_ord}{Integer class labels.}

\item{m}{Number of classes.}

\item{tau_bound}{Absolute bound.}
}
\value{
Updated threshold vector.
}
\description{
Samples new thresholds conditional on the current latent \code{u} and
class labels \code{c_ord}.
}
