% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Study_1_Helper.R
\name{update_u_ess_block_1d}
\alias{update_u_ess_block_1d}
\title{Block elliptical slice sampling update for latent u}
\usage{
update_u_ess_block_1d(
  y,
  u,
  Dx,
  c_ord,
  tau,
  logtheta,
  sigma2_eps,
  block_idx,
  max_try = 300L
)
}
\arguments{
\item{y}{Response vector.}

\item{u}{Current latent vector.}

\item{Dx}{Squared distance matrix.}

\item{c_ord}{Class labels.}

\item{tau}{Thresholds.}

\item{logtheta}{Log GP hyperparameters.}

\item{sigma2_eps}{Error variance.}

\item{block_idx}{Indices of the block to update.}

\item{max_try}{Maximum number of angle proposals.}
}
\value{
List with new \code{u}, number of likelihood evaluations,
and acceptance flag.
}
\description{
Updates a block of latent \code{u} values using elliptical slice sampling
that respects the GP likelihood and ordinal constraints.
}
