% Generated by roxygen2: do not edit by hand
% Please edit documentation in eivGP-literate-source.Rmd.
\docType{package}
\name{eivGP-package}
\alias{eivGP}
\alias{eivGP-package}
\title{eivGP: Errors-in-variables Gaussian processes for ordinal mixed inputs}
\description{
Fits errors-in-variables Gaussian process regressions with numeric predictors, ordinal proxies for latent continuous inputs, and sparse calibration measurements. Includes reproducible Study I and Study II simulation workflows and audited mixed-input GP competitors.
}
\author{
\strong{Maintainer}: Sheng Jiang \email{jiangsheng@cuhk.edu.cn}

Authors:
\itemize{
  \item Sheng Jiang \email{jiangsheng@cuhk.edu.cn}
  \item Penghui Fu
  \item Xinye Chen
  \item Jianhua Z. Huang
}

}
\keyword{internal}