% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eivGP-package.R
\docType{package}
\name{eivGP-package}
\alias{eivGP}
\alias{eivGP-package}
\title{eivGP}
\description{
This package implements the \code{eivGP} method.
}
\author{
\strong{Maintainer}: Sheng Jiang \email{jiangsheng@cuhk.edu.cn}

Authors:
\itemize{
  \item Sheng Jiang \email{jiangsheng@cuhk.edu.cn}
}

}
\keyword{internal}
